+++ consumer/app/build.gradle
@@ -83,6 +83,8 @@ dependencies {
     implementation 'com.github.bumptech.glide:glide:4.9.0'
     kapt 'com.github.bumptech.glide:compiler:4.9.0'
 
+    implementation "androidx.swiperefreshlayout:swiperefreshlayout:1.1.0"
+
     // Recycler View
     implementation 'androidx.recyclerview:recyclerview:1.2.0-alpha05'
