+++ consumer/app/src/main/java/com/videodac/hls/App.kt
@@ -2,8 +2,6 @@ package com.videodac.hls
 
 import androidx.multidex.MultiDexApplication
 
-import org.bouncycastle.jce.provider.BouncyCastleProvider
-import java.security.Security
 
 import com.videodac.hls.helpers.GasOracleHelper.gasOracle
 import com.videodac.hls.helpers.RetrofitHelper
@@ -16,9 +14,6 @@ class App : MultiDexApplication() {
 
     override fun onCreate() {
         super.onCreate()
-
-        // setup the correct crypto lib
-        setupBouncyCastle()
 
         // init web3
         web3 = Utils.getWeb3(this)
@@ -36,25 +31,9 @@ class App : MultiDexApplication() {
         status = retrofitHelper.getStatusService()
     }
 
-    private fun setupBouncyCastle() {
-        val provider =
-            Security.getProvider(BouncyCastleProvider.PROVIDER_NAME)
-                ?: // Web3j will set up the provider lazily when it's first used.
-                return
-        when (provider.javaClass) {
-            BouncyCastleProvider::class.java -> { // BC with same package name, shouldn't happen in real life.
-                return
-            }
-            // Android registers its own BC provider. As it might be outdated and might not include
-            // all needed ciphers, we substitute it with a known BC bundled in the app.
-            // Android's BC has its package rewritten to "com.android.org.bouncycastle" and because
-            // of that it's possible to have another BC implementation loaded in VM.
-            else -> {
-                Security.removeProvider(BouncyCastleProvider.PROVIDER_NAME)
-                Security.insertProviderAt(BouncyCastleProvider(), 1)
-            }
-        }
-    }
+
+
+
 
 
 }
