+++ consumer/app/src/main/java/com/videodac/hls/VideoDacGlideModule.kt
@@ -1,13 +1,14 @@
 package com.videodac.hls
 
 import android.content.Context
-
 import com.bumptech.glide.GlideBuilder
+
 import com.bumptech.glide.annotation.GlideModule
 import com.bumptech.glide.load.engine.DiskCacheStrategy
 import com.bumptech.glide.module.AppGlideModule
 import com.bumptech.glide.request.RequestOptions
 import com.bumptech.glide.signature.ObjectKey
+
 
 @GlideModule
 class VideoDacGlideModule:AppGlideModule() {
@@ -16,4 +17,6 @@ class VideoDacGlideModule:AppGlideModule() {
         super.applyOptions(context, builder)
         builder.apply { RequestOptions().diskCacheStrategy(DiskCacheStrategy.ALL).signature(ObjectKey(System.currentTimeMillis().toShort()))}
     }
+
+
 }
