+++ consumer/app/src/main/res/layout/channel_item.xml
@@ -9,19 +9,25 @@
     android:clickable="true"
     android:focusable="true">
 
+    <com.github.ybq.android.spinkit.SpinKitView
+        android:id="@+id/img_loader"
+        style="@style/SpinKitView.CubeGrid"
+        android:layout_width="90dp"
+        android:layout_height="90dp"
+        android:visibility="gone"
+        android:scaleType="centerCrop"
+        app:SpinKit_Color="@android:color/white" />
 
     <com.lelloman.identicon.view.GithubIdenticonView
         android:id="@+id/channel_identicon"
         android:layout_width="100dp"
-        android:layout_height="100dp"
-        app:hash="654984321" />
+        android:layout_height="100dp" />
 
     <ImageView
         android:id="@+id/channel_icon"
         android:layout_width="100dp"
         android:layout_height="100dp"
         android:contentDescription="@string/channel_icon_desc" />
-
 
     <TextView
         android:id="@+id/channel_address"
